requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
autopep8==2.0.2
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
Flask==2.2.3
idna==3.4
importlib-metadata==6.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
pycodestyle==2.10.0
python-dotenv==0.21.1
pyuwsgi==2.0.21
redis==4.5.4
requests==2.28.2
soupsieve==2.4
tomli==2.0.1
typing-extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.2.3
zipp==3.15.0
